# Conflicts:
#	src/main/java/com/example/scam_detection/entity/EmailRecord.java

+++ src/main/java/com/example/scam_detection/repository/ScamRepository.java
@@ -1,9 +1,7 @@
 package com.example.scam_detection.repository;
 
+import com.example.scam_detection.entity.Scam;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.scam_detection.entity.Scam;
-
-public interface ScamRepository extends JpaRepository<Scam, Integer>{
-    
+public interface ScamRepository extends JpaRepository<Scam,Long> {
 }
